MainScreen_Filter
MainScreen 중간의 Filter 아이콘을 누르면 이미지 <-> 리스트 변경
