+++ src/data/hook/setMainView.js
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import AppStateContext from '../context/AppStateContext';
+
+export default function setMainView() {
+  const { viewImageList, setViewImageList } = useContext(AppStateContext);
+
+  return { viewImageList, setViewImageList };
+}
